+++ changelog.md
@@ -9,6 +9,7 @@
 - Update required Gunfig version to 1.1.11 for stability and bugfixes
 - Allowed the pizza minigame to be initiated without having to find or visit any secret rooms
 - Fixed Entropynnium being completely nonfunctional against Lich's 2nd phase
+- Prevented mastered Alien Nailgun from auto-spawning a few enemies that do more harm than good (e.g., enemies that summon other enemies)
 
 ## 1.32.4 (2026-07-05)